refactor(daemon): expose modules `queries` and `daemon`
The `queries` module exposes (and groups) the structures needed to interact with
a Zenoh-Flow daemon from the outside.

The `daemon` module exposes (and groups) the structure needed to start a
Zenoh-Flow daemon.
