+++ zenoh-flow-daemon/src/instances/queries.rs
@@ -13,7 +13,7 @@
 //
 
 use super::{abort, create, delete, start};
-use crate::Origin;
+use crate::instances::Origin;
 
 use std::{fmt::Debug, sync::Arc};
